+++ anyblok_pyramid/test_bloks/test2/views.py
@@ -10,19 +10,20 @@ class Bloks:
 
     @view_config(route_name='bloks', renderer="json",
                  request_method='GET', permission="read")
-    def get_all(self, request):
-        return self.registry.System.Blok.query().all().to_dict()
+    def get_all(self):
+        return self.registry.System.Blok.query().all().to_dict(
+                'name', 'author', 'version')
 
     @view_config(route_name='blok', renderer="json",
                  request_method='GET', permission="read")
-    def get_one(self, request):
+    def get_one(self):
         return self.registry.System.Blok.query().filter_by(
-            name=self.request.matchdict['name']).one().to_dict()
+            name=self.request.matchdict['name']).one().to_dict(
+                'name', 'author', 'version')
 
     @view_config(route_name='blok', renderer="json",
                  request_method='PUT', permission="write")
-    def put_one(self, request):
-        blok = self.registry.System.Blok.query().filter_by(
-            name=self.request.matchdict['name']).one()
-        blok.update(self.request.json_body)
-        return blok.to_dict()
+    def put_one(self):
+        return self.registry.System.Blok.query().filter_by(
+            name=self.request.matchdict['name']).one().to_dict(
+                'name', 'author', 'version')