+++ anyblok_pyramid/tests/test_blok.py
@@ -24,12 +24,21 @@ class TestPyramidBlok(PyramidDBTestCase):
         registry.upgrade(install=('test-pyramid2',))
         self.webserver.get('/bloks', status=403)
         self.webserver.get('/blok/auth', status=403)
-        self.webserver.post('/login', {'login': 'viewer', 'password': ''},
-                            status=302)
-        self.webserver.get('/bloks', status=200)
-        self.webserver.get('/blok/auth', status=403)
-        self.webserver.post('/logout', {}, status=302)
-        self.webserver.post('/login', {'login': 'admin', 'password': ''},
-                            status=302)
-        self.webserver.get('/bloks', status=200)
-        self.webserver.get('/blok/auth', status=200)
+        resp = self.webserver.post(
+            '/login', {'login': 'viewer', 'password': ''},
+            status=302)
+        headers = resp.headers
+        self.webserver.get('/bloks', status=200, headers=headers)
+        self.webserver.get('/blok/auth', status=200, headers=headers)
+        self.webserver.put('/blok/auth', {}, status=403, headers=headers)
+        resp = self.webserver.post('/logout', {}, status=302)
+        headers = resp.headers
+        self.webserver.get('/bloks', status=403, headers=headers)
+        self.webserver.get('/blok/auth', status=403, headers=headers)
+        resp = self.webserver.post(
+            '/login', {'login': 'admin', 'password': ''},
+            status=302)
+        headers = resp.headers
+        self.webserver.get('/bloks', status=200, headers=headers)
+        self.webserver.get('/blok/auth', status=200, headers=headers)
+        self.webserver.put('/blok/auth', {}, status=200, headers=headers)