+++ doc/CHANGES.rst
@@ -3,6 +3,7 @@
 ..    Copyright (C) 2015 Jean-Sebastien SUZANNE <jssuzanne@anybox.fr>
 ..    Copyright (C) 2016 Jean-Sebastien SUZANNE <jssuzanne@anybox.fr>
 ..    Copyright (C) 2017 Jean-Sebastien SUZANNE <jssuzanne@anybox.fr>
+..    Copyright (C) 2018 Jean-Sebastien SUZANNE <jssuzanne@anybox.fr>
 ..
 .. This Source Code Form is subject to the terms of the Mozilla Public License,
 .. v. 2.0. If a copy of the MPL was not distributed with this file,You can
@@ -16,6 +17,7 @@ CHANGELOG
 0.9.0
 -----
 
+* [FIX] commited session with pyramid
 * [ADD] Authentication configuration
 
 0.8.2 (2017-12-23)