+++ doc/CHANGES.rst
@@ -20,6 +20,7 @@ CHANGELOG
 * [FIX] commited session with pyramid
 * [ADD] Authentication configuration
 * [ADD] **auth** blok
+* [ADD] **auth-password** blok
 
 0.8.2 (2017-12-23)
 ------------------